Uptime, More cog spliting and more updates
