+++ utils/checks.py
@@ -10,6 +10,7 @@ def notblacklisted(ctx):
         return True
 
 async def plugin_enabled(ctx):
+    categories = ["Counting", "Fun", "Leveling", "Logs", "Minecraft", "Moderation", "Music", "Ping", "Search", 'Settings', "Text", "Ticket", "Utilities", "Voice", "Welcome", "Economy", "Games"]
     if ctx.cog is None:
         return True
     with open('./database/db.json') as f:
